move event payloads to separate file to resolve circular dependencies
