+++ src/handlers/gestures/panGesture.ts
@@ -1,9 +1,7 @@
 import { BaseGestureConfig, ContinousBaseGesture } from './gesture';
 import { GestureUpdateEvent } from '../gestureHandlerCommon';
-import {
-  PanGestureConfig,
-  PanGestureHandlerEventPayload,
-} from '../PanGestureHandler';
+import { PanGestureConfig } from '../PanGestureHandler';
+import { PanGestureHandlerEventPayload } from '../GestureHandlerEventPayload';
 
 export type PanGestureChangeEventPayload = {
   changeX: number;
